# swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "version" : "0.0.1",
    "title" : "soeofficer",
    "termsOfService" : "Please read on https://nordkern.dk/wp-content/uploads/2018/03/Projektplan.pdf",
    "contact" : {
      "name" : "Nordkern ApS",
      "url" : "https://nordkern.dk",
      "email" : "kontakt@nordkern.dk"
    }
  },
  "basePath" : "/api",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/officer/{id}" : {
      "get" : {
        "tags" : [ "officer" ],
        "summary" : "Get specific officer from id",
        "description" : "",
        "operationId" : "readOfficer",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Officer"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "officer" ],
        "summary" : "Update an existing officer of the system",
        "description" : "",
        "operationId" : "updateOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "appointedNumber",
          "in" : "query",
          "description" : "The appointed number of the officer",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "dodabNumber",
          "in" : "query",
          "description" : "The officer's Dodab number",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "appointedDate",
          "in" : "query",
          "description" : "The date from which the officer title is valid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "appointedUntil",
          "in" : "query",
          "description" : "The date from which the officer title is invalid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "terminationCause",
          "in" : "query",
          "description" : "The cause of termination",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "person",
          "in" : "query",
          "required" : false,
          "type" : "object"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Officer"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "officer" ],
        "summary" : "Delete an officer from the system",
        "description" : "",
        "operationId" : "deleteOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/promotion" : {
      "post" : {
        "tags" : [ "promotion" ],
        "summary" : "Add a promotion to the system",
        "description" : "",
        "operationId" : "createPromotion",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "dateOfPromotion",
          "in" : "query",
          "description" : "The date from which the rank is valid",
          "required" : true,
          "type" : "date"
        }, {
          "name" : "rank",
          "in" : "query",
          "description" : "The date from which the rank is invalid",
          "required" : true,
          "type" : "object"
        }, {
          "name" : "officer",
          "in" : "query",
          "description" : "The id of the officer possessing the rank",
          "required" : true,
          "type" : "object"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Promotion"
            }
          }
        }
      }
    },
    "/promotion/{id}" : {
      "put" : {
        "tags" : [ "promotion" ],
        "summary" : "Update a promotion of the system",
        "description" : "",
        "operationId" : "updatePromotion",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "dateOfPromotion",
          "in" : "query",
          "description" : "The date from which the rank is valid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "rank",
          "in" : "query",
          "description" : "The date from which the rank is invalid",
          "required" : false,
          "type" : "object"
        }, {
          "name" : "officer",
          "in" : "query",
          "description" : "The id of the officer possessing the rank",
          "required" : false,
          "type" : "object"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Promotion"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "promotion" ],
        "summary" : "Delete a promotion from the system",
        "description" : "",
        "operationId" : "deleteOfficer",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/rank" : {
      "post" : {
        "tags" : [ "rank" ],
        "summary" : "Add a rank to the system",
        "description" : "",
        "operationId" : "createRank",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "natoCode",
          "in" : "query",
          "description" : "The nato code associated with the specific rank",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "rankName",
          "in" : "query",
          "description" : "The name of the rank",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "rankValidFrom",
          "in" : "query",
          "description" : "The date from which the rank is valid",
          "required" : true,
          "type" : "date"
        }, {
          "name" : "rankValidUntil",
          "in" : "query",
          "description" : "The date from which the rank is invalid",
          "required" : true,
          "type" : "date"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rank"
            }
          }
        }
      }
    },
...
    "/rank/all/active" : {
      "post" : {
        "tags" : [ "rank" ],
        "summary" : "Get all rank's active on the specified date",
        "description" : "",
        "operationId" : "findAllActiveOnDate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "date",
          "in" : "query",
          "description" : "The date search criteria",
          "required" : true,
          "type" : "date"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/rank/{id}" : {
      "get" : {
        "tags" : [ "rank" ],
        "summary" : "Get specific rank from id",
        "description" : "",
        "operationId" : "readRank",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rank"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "rank" ],
        "summary" : "Update an existing rank of the system",
        "description" : "",
        "operationId" : "updateRank",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "natoCode",
          "in" : "query",
          "description" : "The nato code associated with the specific rank",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "rankName",
          "in" : "query",
          "description" : "The name of the rank",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "rankValidFrom",
          "in" : "query",
          "description" : "The date from which the rank is valid",
          "required" : false,
          "type" : "date"
        }, {
          "name" : "rankValidUntil",
          "in" : "query",
          "description" : "The date from which the rank is invalid",
          "required" : false,
          "type" : "date"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Rank"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "rank" ],
        "summary" : "Delete a rank from the system",
        "description" : "",
        "operationId" : "deleteRank",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Officer" : {
      "type" : "object",
      "required" : [ "appointedNumber", "dodabNumber", "id", "person", "terminationCause" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the officer"
        },
        "appointedNumber" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 12,
          "description" : "The appointed number of the officer"
        },
        "dodabNumber" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 123,
          "description" : "The officer's Dodab number"
        },
        "appointedDate" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the officer title is valid",
          "readOnly" : true
        },
        "appointedUntil" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the officer title is invalid"
        },
        "terminationCause" : {
          "type" : "string",
          "example" : "Accidental_death",
          "description" : "The cause of termination",
          "enum" : [ "Terminated", "Killed_in_action", "Accidental_death", "Transferred", "Other" ]
        },
        "person" : {
          "description" : "The person associated with the officer",
          "$ref" : "#/definitions/Person"
        }
      }
    },
...
    "Person" : {
      "type" : "object",
      "required" : [ "dateOfBirth", "gender", "givenName", "id", "surname" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the person"
        },
        "givenName" : {
          "type" : "string",
          "example" : "John",
          "description" : "The given name of the person"
        },
        "surname" : {
          "type" : "string",
          "example" : "Doe",
          "description" : "The surname of the person"
        },
        "dateOfBirth" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The person's date of birth. Format is: DD/MM/YYYY"
        },
        "gender" : {
          "type" : "string",
          "example" : "Mand",
          "description" : "The gender of the person. Accepted values are from the set {Mand,Kvinde,Ukendt}",
          "enum" : [ "Male", "Female", "Unknown" ]
        },
        "dateOfDeath" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date of death. Format is: DD/MM/YYYY"
        }
      }
    },
...
    "Promotion" : {
      "type" : "object",
      "required" : [ "dateOfPromotion", "id", "rank" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the promotion"
        },
        "dateOfPromotion" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "1",
          "description" : "The date from which the rank is valid"
        },
        "rank" : {
          "example" : "1",
          "description" : "The date from which the rank is invalid",
          "$ref" : "#/definitions/Rank"
        }
      }
    },
    "Rank" : {
      "type" : "object",
      "required" : [ "id", "natoCode", "rankName", "rankNameEn", "rankValidFrom", "rankValidUntil" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1,
          "description" : "The unique ID of the rank"
        },
        "natoCode" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 1234,
          "description" : "The nato code of the rank"
        },
        "rankName" : {
          "type" : "string",
          "example" : "Officer",
          "description" : "The name of the rank"
        },
        "rankNameEn" : {
          "type" : "string",
          "example" : "Officer",
          "description" : "The english name of the rank"
        },
        "rankValidFrom" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the rank is valid"
        },
        "rankValidUntil" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "01/01/2018",
          "description" : "The date from which the rank is invalid"
        }
      }
    },
...
  }
}